requirements.txt
asyncpg==0.30.0
discord.py==2.5.2
python-dotenv==1.1.1
